pyproject.toml
[project]
name = "unifi-camera-manager"
version = "0.2.0"
description = "CLI tool to manage UniFi Protect cameras including third-party ONVIF cameras"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "httpx>=0.28.0",
    "onvif-zeep-async>=4.0.4",
    "platformdirs>=4.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.2",
    "rich>=14.2.0",
    "typer[all]>=0.21.1",
    "uiprotect>=8.0.0",
]

[project.scripts]
ucam = "unifi_camera_manager.cli:main"

[project.optional-dependencies]
dev = [
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unifi_camera_manager"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "D"]
ignore = ["D100", "D104", "D107"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.coverage.run]
source = ["src/unifi_camera_manager"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
